refactor: merge Loki and Prometheus integration tests
Combine the two separate test modules into a single
test_observability.py that deploys k6 alongside both Loki and
Prometheus, then verifies logs and metrics after each run.

A single load_test.js script replaces the two endpoint-specific
scripts; it generates both console.log entries (for Loki) and HTTP
metrics (for Prometheus).
