+++ tests/integration/test_observability.py
@@ -1,7 +1,7 @@
 # Copyright 2025 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-"""Integration tests for k6-k8s with Prometheus remote write."""
+"""Integration tests for k6-k8s with Loki log shipping and Prometheus remote write."""
 
 import json
 import time
@@ -11,8 +11,9 @@ import pytest
 
 from .conftest import APP_NAME, K6_IMAGE, RESOURCES_DIR
 
+LOKI_APP = "loki"
 PROMETHEUS_APP = "prometheus-k8s"
-LOAD_TEST_SCRIPT = (RESOURCES_DIR / "load_test_prometheus.js").read_text()
+LOAD_TEST_SCRIPT = (RESOURCES_DIR / "load_test.js").read_text()
 
 
 def _wait_for_idle(juju: jubilant.Juju, timeout=300, poll_interval=10):
@@ -27,9 +28,38 @@ def _wait_for_idle(juju: jubilant.Juju, timeout=300, poll_interval=10):
     raise TimeoutError(f"{APP_NAME} did not become idle within {timeout}s")
 
 
+def _assert_loki_has_k6_logs(juju: jubilant.Juju, retries=30, delay=10):
+    """Query Loki for k6 log entries, retrying until found."""
+    now = int(time.time())
+    start = now - 600  # look back 10 minutes
+    end = now + 60
+    query = '{' + f'juju_application="{APP_NAME}"' + '}'
+    for _ in range(retries):
+        try:
+            output = juju.ssh(
+                f"{LOKI_APP}/0",
+                "curl",
+                "-sG",
+                "http://localhost:3100/loki/api/v1/query_range",
+                "--data-urlencode", f"query={query}",
+                "-d", f"start={start}",
+                "-d", f"end={end}",
+                "-d", "limit=10",
+                container="loki",
+            )
+            data = json.loads(output)
+            results = data.get("data", {}).get("result", [])
+            if results:
+                return
+        except Exception:
+            pass
+        time.sleep(delay)
+    raise AssertionError("No k6 logs found in Loki")
+
+
 def _assert_prometheus_has_k6_metrics(juju: jubilant.Juju, retries=30, delay=10):
     """Query Prometheus for k6 metrics, retrying until found."""
-    for attempt in range(retries):
+    for _ in range(retries):
         try:
             output = juju.ssh(
                 f"{PROMETHEUS_APP}/0",
@@ -48,26 +78,28 @@ def _assert_prometheus_has_k6_metrics(juju: jubilant.Juju, retries=30, delay=10)
 
 
 # ---------------------------------------------------------------------------
-# Single k6 unit + Prometheus
+# Single k6 unit + Loki + Prometheus
 # ---------------------------------------------------------------------------
 
 
 @pytest.mark.juju_setup
 @pytest.mark.abort_on_fail
-def test_deploy_prometheus(juju: jubilant.Juju, charm_path):
-    """Deploy k6 and prometheus, integrate them."""
+def test_deploy(juju: jubilant.Juju, charm_path):
+    """Deploy k6, Loki and Prometheus, then integrate them."""
     juju.deploy(charm_path, APP_NAME, resources={"k6-image": K6_IMAGE})
+    juju.deploy(LOKI_APP, channel="dev/edge", trust=True)
     juju.deploy(PROMETHEUS_APP, channel="dev/edge", trust=True)
     juju.wait(
-        lambda s: jubilant.all_active(s, APP_NAME, PROMETHEUS_APP),
+        lambda s: jubilant.all_active(s, APP_NAME, LOKI_APP, PROMETHEUS_APP),
         timeout=600,
     )
+    juju.integrate(f"{APP_NAME}:logging", LOKI_APP)
     juju.integrate(f"{APP_NAME}:send-remote-write", PROMETHEUS_APP)
     juju.wait(lambda s: jubilant.all_active(s), timeout=300)
 
 
 def test_run_single_unit(juju: jubilant.Juju):
-    """Configure and run a lightweight load test, then verify Prometheus has k6 metrics."""
+    """Configure and run a load test, then verify both Loki logs and Prometheus metrics."""
     juju.config(APP_NAME, values={"load-test": LOAD_TEST_SCRIPT})
     juju.wait(lambda s: jubilant.all_active(s), timeout=120)
 
@@ -75,16 +107,17 @@ def test_run_single_unit(juju: jubilant.Juju):
     task.raise_on_failure()
 
     _wait_for_idle(juju)
+    _assert_loki_has_k6_logs(juju)
     _assert_prometheus_has_k6_metrics(juju)
 
 
 # ---------------------------------------------------------------------------
-# Multiple k6 units + Prometheus
+# Multiple k6 units + Loki + Prometheus
 # ---------------------------------------------------------------------------
 
 
 def test_scale_up_and_run_multi_unit(juju: jubilant.Juju):
-    """Scale k6 to 3 units, run load test, verify Prometheus metrics."""
+    """Scale k6 to 3 units, run load test, verify Loki logs and Prometheus metrics."""
     juju.cli("scale-application", APP_NAME, "3")
     juju.wait(lambda s: jubilant.all_active(s, APP_NAME), timeout=600)
 
@@ -92,4 +125,5 @@ def test_scale_up_and_run_multi_unit(juju: jubilant.Juju):
     task.raise_on_failure()
 
     _wait_for_idle(juju)
+    _assert_loki_has_k6_logs(juju)
     _assert_prometheus_has_k6_metrics(juju)
